An argument is an object or an expression passed to a function — added between the opening and closing parentheses — when it is called:
Example Code
greet = 'Hello!'
print(greet)
The code in the example above would print the string 'Hello!', which is the value of the variable greet passed to print() as the argument.

Print your text variable to the screen by passing the text variable as the argument to the print() function.