You are going to use the .find() method to find the position in the alphabet of each letter in your message. A method is similar to a function, but it belongs to an object.
Example Code
sentence = 'My brain hurts!'
sentence.find('r')
Above, the .find() method is called on sentence (the string to search in), and 'r' (the character to locate) is passed as the argument. The sentence.find('r') call will return 4, which is the index of the first occurrence of 'r' in sentence.

At the end of your code, call .find() on alphabet and pass 'z' as the argument to the method.